+++ docs/CHANGELOG.md
@@ -10,7 +10,7 @@ Focus this version was on performance and security by default.
 
 ## Breaking Changes
 - Updated `aws-sdk` to v3
-- Updated all packages to be ES6 modules
+- Updated all packages to be ES6 modules (esm)
 - All middlewares now use `async/await` and have deprecated `next(err)` and `callback(err, response)`
 - `validator` refactored to support `draft-2019-09` using the latest version of `ajv`. Full `i18n` is now enabled by default (MAYBE)
 - Middlewares that reach out to 3rd party API have been completely refactored to have unifying options that resolve on demand from internal context. Applies to: